Avoid read timeout when polling after a sequence
After having taken a sequence of images, we need to call pvc_get_frame to pull the data out of the buffer into python memory. As written, each call to this function was waiting an extra 200ms unnecessarily. This is because, when we call pvc_get_frame in this situation, unlike when pvc_get_frame is used in live mode, there will never be a new frame. Rather, we are calling this function to get frames we've already captured.

To avoid the 200ms timeout, check whether there is existing data at the start of the while loop.