Fixed minor error reporting issue in bin_y setter
Incorrect error message might have been generated by the exception in the bin_y setter because PARAM_BINNING_SER was used to print the array of allowed bin-y values in the error message. This was most likely a typo, PARAM_BINNING_PAR should be used.
This was not a functional error, it might have been, however, confusing for people reading the code or interpreting the possible error message.