Reworked locking in PVC module
The global camera instance map holds smart pointers with Camera instances and the global mutex guards access to that map only.
Once we have the Camera instance from the map, we lock the Camera::m_mutex that remains locked all the time the Camera instance is needed. It includes get_frame function that unlocks it only while passively waiting for new frame.